Refactor pantry components
