+++ client/KitchenFox/components/app.js
@@ -4,7 +4,7 @@ import SigninContainer from './auth/signin_container';
 import Greeting from './auth/greeting';
 import { checkLogin } from '../actions/session_actions';
 
-import Pantry from '../screens/pantry_category_index';
+import Pantry from '../screens/pantry';
 
 class App extends Component {
   constructor(props) {
